Modified metrics definition
updatedAt attribute in K8sDeploymentSample sounds as a timestamp, but it
actually means how many replicas have been updated updated at the deployment
deployment (taken from the kube_deployment_status_replicas_updated metric).
So the metric has been renamed to `podsUpdated`.

cpuUsedCores shows the number of used nanocores, so it reports a very big
number which may be confusing. To solve that, the metric will be divided
by 10^9.

Signed-off-by: Mario Macias <mmacias@newrelic.com>