Kubernetes: Return proper error on KSM/Kubelet connection error
The Group() method now returns an ErrorGroup instance instead of
a slice of errors.

The ErrorGroup instance has a field that states whether the received
error is recoverable or not. So in non-recoverable errors (e.g.
connection failure to KSM or Kubelet) the execution of the plugin
stops and for recoverable errors (e.g. a single metric entry can't
be parsed), the errors are just recorded and later logged.

Signed-off-by: Mario Macias <mmacias@newrelic.com>