Update e2e/jsonschema to validate missing event types
If our integration doesn't return all the expected event types that means that
it isn't working properly. With this change we validate that all the event types
that we pass in a map to the `match` function are present in the output of our
integration.

I've updated also how we print the error messages in order to make it more
readable. Ex:

```
=== RUN   TestBasic
Executing our integration in "https://192.168.99.100:8443" cluster. K8s version: v1.5.3
--- FAIL: TestBasic (0.44s)
        Error Trace:    k8s_test.go:162
	Error:      	Should be empty, but was [
	            	------ leader pod newrelic-infra-4qwqn ------

	            		- Mandatory types were not found: K8sPodSample, K8sContainerSample, K8sNodeSample,
	            		- Some types were not validated because no schema was found: K8sReplicasetSample, K8sNamespaceSample, K8sDeploymentSample,
	            	]
	Test:       	TestBasic
FAIL
```