Fix fsUsedPercent calculation
As documented in the Kubernetes source code at https://github.com/kubernetes/kubernetes/blob/master/pkg/volume/volume.go#L70,
the `fsCapacityBytes` metrics might not be equal to `fsUsedBytes` + `fsAvailableBytes` in some scenarios.
So we consider the total bytes in a storage as the used + available amounts, as when available gets close to 0 the container
won't be able to write to that storage, even if the capacity - used still isn't zero.
.
Then the `fsUsedPercent` is calculated as `fsUsedBytes / (fsAvailableBytes + fsUsedBytes)`.