Update Makefile to use specific versions of our tooling (#6)
* Update Makefile to use specific version of our tooling

The previous version of this Makefile doesn't use pinned versions of our
tooling. That means that each developer of the project could end up
having different versions of the tooling, as well as Jenkins, and we
could have weird problems.

To solve this, I've updated the Makefile to download specific versions
for `golangci-lint` and `govendor`. For `papers-go` isn't possible to
follow the same approach because the repository isn't public and it
isn't possible to download the binary file without being logged in.

In addition to this, I've cleaned up a little bit the Makefile, this a
summary of the changes:

- Use specific versions for `govendor` and `golangci-lint`
- Simplify the output messages
- Ensure each phony target is marked as phony
- Remove go-cov dependency. In general, we aren't using it. In case
someone wants to use it, they could just execute the command manually
- Remove repetitive targets (ex: validate)