Add a nonprivileged version of our daemonset (#25)
* Add a nonprivileged version of our daemonset

To support environments with high security concerns we would like to
provide a version of our integration that doesn't rely on having
privileged permissions.

* Rename docker image and add a label to the k8s objects

This label could help us to identify if the daemonset is installed in
nonprivileged mode.

* Replace `nonprivileged` by `unprivileged`

Unprivileged word is used by the Infrastructure Agent and I'd like to be
consistent with this.

* Remove the custom docker file for unprivileged version

Instead of having a different dockerfile, replace it by using build
arguments and multi-stage in our main dockerfile. Having different
dockerfiles is harder to mantain. Also, once the new versions of the
base image are released, the different stages won't be necessary and we
will need to use only a different image tag.

* Update paths in the manifest file for unprivileged version

As we are mounting them as read only, we can use the default paths for
the Infrastructure Agent.

* Update image name in the manifest file for unprivileged version

Instead of expecting a different repository/image, what we will do is
publishing the unprivileged version as a different tag.

* Update skaffold changing the tagPolicy

Using as a tagPolicy the sha256 of the docker images we could use the
same repository for unprivileged an normal versions.

* Update Jenkinsfile to build the unprivileged image

Also, change how the image is built by Jenkins. There is a known issue
with the Docker Jenkins plugin and docker multi stages. Ref: https://issues.jenkins-ci.org/browse/JENKINS-44609

* Update Jenkinsfile with build-args for unprivileged image

* Fix Jenkinsfile