Update Jenkinsfile to execute e2e tests for unprivileged version (#27)
* Update Jenkinsfile to execute e2e tests for unprivileged version

Execute them in parallel in two different clusters. We have decided two
use two different clusters with two different k8s versions because at
this moment we don't have a mechanism to test against all the k8s
versions in a dynamic way.

Note: The `fsi-jenkins-112-rbac` cluster doesn't exist yet.

* Rename cluster and add a different cluster domain

New clusters aren't created in `aws.datanerd.us` domain anymore.

* Fix cluster name value

* Execute E2E tests for unprivileged version in the existing cluster

We don't have the new cluster yet and we would like to avoid being
blocked. The build will take more time to pass but we will address that
in the future.

* Update E2E tests and chart templates

For the unprivileged version we need different values in our daemonset.
To test this properly in our E2E tests, we need to update the helm chart
template to be able to differenciate between the unprivileged version
and the normal one.

* Fix string format

* Pass unprivileged arg to the E2E tests

* Fix cli arg for unprivileged mode