Optional Distributed KSM parsing (#91)
**How does this even work?**

At a high level, customers will:

1. Enable the feature using an environment variable
2. Use another environment variable to tell which label to use to find KSM instances (will be referred to later as "KSM pod label")
3. Deploy multiple KSM, each one with different collectors enabled, and running on different nodes to ensure maximum load distribution. This can be changed given what the customer things is acceptable regarding the load.

Then, the behaviour of the integration in regards to the discovery of KSM will follow these steps:

1. Each integration pod will query the Kubernetes API for pods that contain the KSM pod label.
2. Will filter out these results by matching its own NodeIP against the pods found. 
3. Will keep a list of HTTP clients to acess all these KSM instances.
4. Each integration pod will query **all** KSM instances in its node matching the KSM pod label.
5. Profit.

**Some technical details**

Adapting the current discovery cache mechanism to return a slice of clients instead of a single clients is crazy and too much work. I actually tried this for a while.

Extending the discovery mechanism by adding this discovery that gets multiple clients seemed like a good solution that could be done in a relatively reasonable time. It also doesn't add a lot of complexity and it not huge.

**Non-goals:**

* (developer) documentation about how to use this feature 
* E2E tests on top of this feature (requires A LOT of refactor)
* Fixing annoyances that are present in using this feature (i.e. the spam of log messages about not being able to find all the KSM data in the instances of the integration that are querying KSM)