Fix bug were the service selectors weren't added to the sample. (#98)
This bug was caused when the randomly selected metric for
constructing the raw entity ID to be used for retrieving the
labels didn't contain the namespace and service labels.

This happened in the case when the randomly selected metric was
the one with the selectors. Since that metric doesn't come from
prometheus, instead we created it from the apiserver services
spec, we weren't adding the required attributes.

Now instead of selecting a random metric when generating the raw
entity id for the metric we try to get a random metric that contains
the required labels.